requirements.txt
censys-platform
python-dotenv
mcp
